refactor(comment): separate comment and child comment into smaller components to avoid duplicates
