+++ src/components/custom/comment-section/comment-details.tsx
@@ -1,0 +1,24 @@
+import ReplySection from "@/components/custom/comment-section/reply-section/reply-section"
+import useRelativeTime from "@/hooks/useRelativeTime"
+import { ChildCommentType, CommentType } from "@/validators/schemas/nfts/comments/commentSchema"
+
+export default function CommentDetails({ comment, nftId, parentId }: CommentDetailsProps) {
+  const relativeTime = useRelativeTime(comment.createdAt)
+
+  return (
+    <div>
+      <h4 className="text-sm font-semibold mb-1">
+        @{comment.author.username}
+        <div className="text-xs">{relativeTime}</div>
+      </h4>
+      <p className="text-md leading-relaxed">{comment.content}</p>
+      <ReplySection nftId={nftId} parentId={parentId} />
+    </div>
+  )
+}
+
+type CommentDetailsProps = {
+  comment: CommentType | ChildCommentType
+  nftId: number
+  parentId: number
+}
